chore: remove useTrackTopic hook, use helper function instead
